Support JAX debug callback on ScaledArrays
Intermediate solution at the moment due to JAX debug callback limitations.
It is not possible at present to get the original callback + input PyTree from
the JAX debug_callback primitive, meaning the interpreter can not modify properly
the graph.

In this PR, we re-implement the high level API `debug_callback` of JAX in order
to save the former, and be able to adapt the call to ScaledArrays.