Add a default scale dtype to AutoScaleConfig
The default scale dtype can be set with `AutoScaleConfig(scale_dtype=xxx)` context manager.
Having a default scale dtype information when the dtype of `data` and `scale` are not the same. If no
information is provided, the `autoscale` JAX interpreter has no way of deciding which dtype is the proper
one during the initial tracing.